Scoring
Added scoring mechanism to the popping of balloons.  The functionallity works but the scoring of points for each different color of balloon should be adjusted to make sense within the game.  This can be easily modified.
Added score update to score counter at bottom of page in game.
Added score update to message displaying that the player wins.
Added score update ot message displaying if the player loses.  Displays current score.